requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
gevent==21.12.0
greenlet==1.1.3
idna==3.4
kombu==5.2.4
lxml==4.9.1
packaging==21.3
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.4
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
